refactor: exception 형태 변경
