+++ backend/src/main/java/consolelog/comment/exception/InvalidMessageException.java
@@ -1,9 +1,9 @@
 package consolelog.comment.exception;
 
-import consolelog.global.advice.BadRequestException;
 import consolelog.global.error.ErrorCode;
+import consolelog.global.exception.BusinessException;
 
-public class InvalidMessageException extends BadRequestException {
+public class InvalidMessageException extends BusinessException {
 
     public InvalidMessageException() {
         super(ErrorCode.INVALID_MESSAGE_ERROR);
