--- backend/src/main/java/consolelog/global/advice/BadRequestException.java
@@ -1,9 +1,0 @@
-package consolelog.global.advice;
-
-import consolelog.global.error.ErrorCode;
-
-public class BadRequestException extends BusinessException {
-    public BadRequestException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
