--- backend/src/main/java/consolelog/global/advice/ForbiddenException.java
@@ -1,10 +1,0 @@
-package consolelog.global.advice;
-
-import consolelog.global.error.ErrorCode;
-
-public class ForbiddenException extends BusinessException {
-
-    public ForbiddenException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
