--- backend/src/main/java/consolelog/global/advice/InternalException.java
@@ -1,9 +1,0 @@
-package consolelog.global.advice;
-
-import consolelog.global.error.ErrorCode;
-
-public class InternalException extends BusinessException {
-    public InternalException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
