--- backend/src/main/java/consolelog/global/advice/NotFoundException.java
@@ -1,10 +1,0 @@
-package consolelog.global.advice;
-
-import consolelog.global.error.ErrorCode;
-
-public class NotFoundException extends BusinessException {
-
-    public NotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
