--- backend/src/main/java/consolelog/global/advice/UnauthorizedException.java
@@ -1,10 +1,0 @@
-package consolelog.global.advice;
-
-import consolelog.global.error.ErrorCode;
-
-public class UnauthorizedException extends BusinessException {
-
-    public UnauthorizedException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
