+++ backend/src/main/java/consolelog/global/error/GlobalExceptionHandler.java
@@ -1,19 +1,13 @@
 package consolelog.global.error;
 
-import consolelog.global.advice.BusinessException;
-import lombok.extern.log4j.Log4j2;
+import consolelog.global.exception.BusinessException;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.Marker;
-import org.apache.logging.log4j.MarkerManager;
 import org.springframework.core.annotation.Order;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.util.Arrays;
-import java.util.stream.Collectors;
 
 @Slf4j
 @RestControllerAdvice
