+++ backend/src/main/java/consolelog/global/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package consolelog.global.advice;
+package consolelog.global.exception;
 
 import consolelog.global.error.ErrorCode;
 import lombok.Getter;
