+++ backend/src/main/java/consolelog/image/exception/S3UploadException.java
@@ -1,6 +1,6 @@
 package consolelog.image.exception;
 
-import consolelog.global.advice.BusinessException;
+import consolelog.global.exception.BusinessException;
 import consolelog.global.error.ErrorCode;
 
 public class S3UploadException extends BusinessException {
