+++ backend/src/main/java/consolelog/member/exception/MemberNotFoundException.java
@@ -1,9 +1,9 @@
 package consolelog.member.exception;
 
-import consolelog.global.advice.NotFoundException;
+import consolelog.global.exception.BusinessException;
 import consolelog.global.error.ErrorCode;
 
-public class MemberNotFoundException extends NotFoundException {
+public class MemberNotFoundException extends BusinessException {
 
     public MemberNotFoundException() {
         super(ErrorCode.MEMBER_NOT_FOUND_ERROR);
