+++ backend/src/main/java/consolelog/post/exception/InvalidContentException.java
@@ -1,9 +1,9 @@
 package consolelog.post.exception;
 
-import consolelog.global.advice.BadRequestException;
 import consolelog.global.error.ErrorCode;
+import consolelog.global.exception.BusinessException;
 
-public class InvalidContentException extends BadRequestException {
+public class InvalidContentException extends BusinessException {
 
     public InvalidContentException() {
         super(ErrorCode.INVALID_CONTENT_ERROR);
