+++ backend/src/main/java/consolelog/post/exception/InvalidTitleException.java
@@ -1,9 +1,9 @@
 package consolelog.post.exception;
 
-import consolelog.global.advice.BadRequestException;
 import consolelog.global.error.ErrorCode;
+import consolelog.global.exception.BusinessException;
 
-public class InvalidTitleException extends BadRequestException {
+public class InvalidTitleException extends BusinessException {
 
     public InvalidTitleException() {
         super(ErrorCode.INVALID_TITLE_ERROR);
