+++ backend/src/main/java/consolelog/post/exception/PostNotFoundException.java
@@ -1,9 +1,9 @@
 package consolelog.post.exception;
 
-import consolelog.global.advice.NotFoundException;
+import consolelog.global.exception.BusinessException;
 import consolelog.global.error.ErrorCode;
 
-public class PostNotFoundException extends NotFoundException {
+public class PostNotFoundException extends BusinessException {
 
     public PostNotFoundException() {
         super(ErrorCode.POST_NOT_FOUND_ERROR);
